refact(rds): Move mock to dedicated package
