+++ internal/app/rds/rds_test.go
@@ -1,8 +1,6 @@
 package rds_test
 
 import (
-	"context"
-	"crypto/rand"
 	"fmt"
 	"reflect"
 	"testing"
@@ -12,100 +10,21 @@ import (
 	aws_rds "github.com/aws/aws-sdk-go-v2/service/rds"
 	aws_rds_types "github.com/aws/aws-sdk-go-v2/service/rds/types"
 	"github.com/qonto/prometheus-rds-exporter/internal/app/rds"
+	mock "github.com/qonto/prometheus-rds-exporter/internal/app/rds/mock"
 	converter "github.com/qonto/prometheus-rds-exporter/internal/app/unit"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-type mockRDSClient struct {
-	DescribeDBInstancesOutput               *aws_rds.DescribeDBInstancesOutput
-	DescribePendingMaintenanceActionsOutput *aws_rds.DescribePendingMaintenanceActionsOutput
-	DescribeDBLogFilesOutput                *aws_rds.DescribeDBLogFilesOutput
-	DescribeDBLogFilesOutputError           error
-	Error                                   error
-}
-
-func (m mockRDSClient) DescribeDBInstancesPages(input *aws_rds.DescribeDBInstancesInput, fn func(*aws_rds.DescribeDBInstancesOutput, bool) bool) error {
-	fn(m.DescribeDBInstancesOutput, false)
-
-	return nil
-}
-
-func (m mockRDSClient) DescribePendingMaintenanceActions(context.Context, *aws_rds.DescribePendingMaintenanceActionsInput, ...func(*aws_rds.Options)) (*aws_rds.DescribePendingMaintenanceActionsOutput, error) {
-	return m.DescribePendingMaintenanceActionsOutput, m.Error
-}
-
-func (m mockRDSClient) DescribeDBLogFiles(ctx context.Context, input *aws_rds.DescribeDBLogFilesInput, fn ...func(*aws_rds.Options)) (*aws_rds.DescribeDBLogFilesOutput, error) {
-	return m.DescribeDBLogFilesOutput, m.DescribeDBLogFilesOutputError
-}
-
-func (m mockRDSClient) DescribeDBInstances(context.Context, *aws_rds.DescribeDBInstancesInput, ...func(*aws_rds.Options)) (*aws_rds.DescribeDBInstancesOutput, error) {
-	return m.DescribeDBInstancesOutput, nil
-}
-
-// randomString returns a random alphanumeric string of the specified length
-func randomString(length int) string {
-	buf := make([]byte, length)
-
-	_, err := rand.Read(buf)
-	if err != nil {
-		panic(err) // out of randomness, should never happen
-	}
-
-	return fmt.Sprintf("%x", buf)
-}
-
-func newRdsCertificateDetails() *aws_rds_types.CertificateDetails {
-	return &aws_rds_types.CertificateDetails{
-		CAIdentifier: aws.String("rds-ca-2019"),
-		ValidTill: aws.Time(time.Date(
-			2024, time.August, 22,
-			17, 8, 50, 0, time.UTC,
-		)),
-	}
-}
-
-func newRdsInstance() *aws_rds_types.DBInstance {
-	awsRegion := "eu-west-3"
-	awsAccountID := "123456789012"
-	DBInstanceIdentifier := randomString(10)
-	arn := fmt.Sprintf("arn:aws:rds:%s:%s:db:%s", awsRegion, awsAccountID, DBInstanceIdentifier)
-
-	now := time.Now()
-
-	return &aws_rds_types.DBInstance{
-		AllocatedStorage:           aws.Int32(5),
-		BackupRetentionPeriod:      aws.Int32(7),
-		DBInstanceArn:              aws.String(arn),
-		DBInstanceClass:            aws.String("t3.large"),
-		DBInstanceIdentifier:       aws.String(DBInstanceIdentifier),
-		DBInstanceStatus:           aws.String("available"),
-		DbiResourceId:              aws.String("resource1"),
-		DeletionProtection:         aws.Bool(true),
-		Engine:                     aws.String("postgres"),
-		EngineVersion:              aws.String("14.9"),
-		Iops:                       aws.Int32(3000),
-		MaxAllocatedStorage:        aws.Int32(10),
-		MultiAZ:                    aws.Bool(true),
-		PerformanceInsightsEnabled: aws.Bool(true),
-		PubliclyAccessible:         aws.Bool(true),
-		StorageType:                aws.String("gp3"),
-		CACertificateIdentifier:    aws.String("rds-ca-2019"),
-		CertificateDetails:         newRdsCertificateDetails(),
-		InstanceCreateTime:         &now,
-		TagList:                    []aws_rds_types.Tag{{Key: aws.String("Environment"), Value: aws.String("unittest")}, {Key: aws.String("Team"), Value: aws.String("sre")}},
-	}
-}
-
 func TestGetMetrics(t *testing.T) {
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{CollectLogsSize: true}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 
@@ -136,23 +55,23 @@ func TestGetMetrics(t *testing.T) {
 }
 
 func TestGP2StorageType(t *testing.T) {
-	rdsInstanceWithSmallDisk := newRdsInstance()
+	rdsInstanceWithSmallDisk := mock.NewRdsInstance()
 	rdsInstanceWithSmallDisk.StorageType = aws.String("gp2")
 	rdsInstanceWithSmallDisk.AllocatedStorage = aws.Int32(10)
 
-	rdsInstanceWithMediumDisk := newRdsInstance()
+	rdsInstanceWithMediumDisk := mock.NewRdsInstance()
 	rdsInstanceWithMediumDisk.StorageType = aws.String("gp2")
 	rdsInstanceWithMediumDisk.AllocatedStorage = aws.Int32(400)
 
-	rdsInstanceWithLargeDisk := newRdsInstance()
+	rdsInstanceWithLargeDisk := mock.NewRdsInstance()
 	rdsInstanceWithLargeDisk.StorageType = aws.String("gp2")
 	rdsInstanceWithLargeDisk.AllocatedStorage = aws.Int32(20000)
 
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstanceWithSmallDisk, *rdsInstanceWithMediumDisk, *rdsInstanceWithLargeDisk}}
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, int64(100), metrics.Instances[*rdsInstanceWithSmallDisk.DBInstanceIdentifier].MaxIops, "Minimum is 100 IOPS")
@@ -166,21 +85,21 @@ func TestGP2StorageType(t *testing.T) {
 }
 
 func TestGP3StorageType(t *testing.T) {
-	rdsInstanceWithSmallDisk := newRdsInstance()
+	rdsInstanceWithSmallDisk := mock.NewRdsInstance()
 	rdsInstanceWithSmallDisk.StorageType = aws.String("gp3")
 	rdsInstanceWithSmallDisk.AllocatedStorage = aws.Int32(10)
 	rdsInstanceWithSmallDisk.Iops = aws.Int32(3000)
 
-	rdsInstanceWithLargeDisk := newRdsInstance()
+	rdsInstanceWithLargeDisk := mock.NewRdsInstance()
 	rdsInstanceWithLargeDisk.StorageType = aws.String("gp3")
 	rdsInstanceWithLargeDisk.AllocatedStorage = aws.Int32(500)
 	rdsInstanceWithLargeDisk.Iops = aws.Int32(12000)
 
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstanceWithSmallDisk, *rdsInstanceWithLargeDisk}}
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, int64(3000), metrics.Instances[*rdsInstanceWithSmallDisk.DBInstanceIdentifier].MaxIops, "IOPS should the same than RDS instance information")
@@ -188,27 +107,27 @@ func TestGP3StorageType(t *testing.T) {
 }
 
 func TestIO1StorageType(t *testing.T) {
-	rdsInstanceWithSmallIOPS := newRdsInstance()
+	rdsInstanceWithSmallIOPS := mock.NewRdsInstance()
 	rdsInstanceWithSmallIOPS.StorageType = aws.String("io1")
 	rdsInstanceWithSmallIOPS.Iops = aws.Int32(1000)
 
-	rdsInstanceWithMediumIOPS := newRdsInstance()
+	rdsInstanceWithMediumIOPS := mock.NewRdsInstance()
 	rdsInstanceWithMediumIOPS.StorageType = aws.String("io1")
 	rdsInstanceWithMediumIOPS.Iops = aws.Int32(4000)
 
-	rdsInstanceWithLargeIOPS := newRdsInstance()
+	rdsInstanceWithLargeIOPS := mock.NewRdsInstance()
 	rdsInstanceWithLargeIOPS.StorageType = aws.String("io1")
 	rdsInstanceWithLargeIOPS.Iops = aws.Int32(48000)
 
-	rdsInstanceWithHighIOPS := newRdsInstance()
+	rdsInstanceWithHighIOPS := mock.NewRdsInstance()
 	rdsInstanceWithHighIOPS.StorageType = aws.String("io1")
 	rdsInstanceWithHighIOPS.Iops = aws.Int32(64000)
 
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstanceWithSmallIOPS, *rdsInstanceWithMediumIOPS, *rdsInstanceWithLargeIOPS, *rdsInstanceWithHighIOPS}}
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, converter.MegaBytesToBytes(int64(250)), metrics.Instances[*rdsInstanceWithSmallIOPS.DBInstanceIdentifier].StorageThroughput, "Minimum is 256 MiB/s")
@@ -219,7 +138,7 @@ func TestIO1StorageType(t *testing.T) {
 
 func TestLogSize(t *testing.T) {
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
 	// Mock log files
@@ -234,13 +153,13 @@ func TestLogSize(t *testing.T) {
 
 	mockDescribeDBLogFilesOutput := &aws_rds.DescribeDBLogFilesOutput{DescribeDBLogFiles: rdsLogFiles}
 
-	mock := mockRDSClient{
+	client := mock.RDSClient{
 		DescribeDBInstancesOutput: mockDescribeDBInstancesOutput,
 		DescribeDBLogFilesOutput:  mockDescribeDBLogFilesOutput,
 	}
 	configuration := rds.Configuration{CollectLogsSize: true}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, aws.Int64(expectedLogFilesSize), metrics.Instances[*rdsInstance.DBInstanceIdentifier].LogFilesSize, "Log files size mismatch")
@@ -248,16 +167,16 @@ func TestLogSize(t *testing.T) {
 
 func TestLogSizeInCreation(t *testing.T) {
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{
+	client := mock.RDSClient{
 		DescribeDBInstancesOutput:     mockDescribeDBInstancesOutput,
 		DescribeDBLogFilesOutputError: &aws_rds_types.DBInstanceNotFoundFault{},
 	}
 	configuration := rds.Configuration{CollectLogsSize: true}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	var emptyInt64 *int64
 
@@ -269,14 +188,14 @@ func TestReplicaNode(t *testing.T) {
 	primaryInstance := "primary-instance"
 
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	rdsInstance.ReadReplicaSourceDBInstanceIdentifier = aws.String(primaryInstance)
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{CollectLogsSize: true}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, "replica", metrics.Instances[*rdsInstance.DBInstanceIdentifier].Role, "Should be replica")
@@ -321,13 +240,13 @@ func TestGetDBInstanceStatusCode(t *testing.T) {
 
 func TestPendingModification(t *testing.T) {
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, false, metrics.Instances[*rdsInstance.DBInstanceIdentifier].PendingModifiedValues, "Should not have any pending modification")
@@ -335,15 +254,15 @@ func TestPendingModification(t *testing.T) {
 
 func TestPendingModificationDueToInstanceModification(t *testing.T) {
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	pendingModifications := aws_rds_types.PendingModifiedValues{AllocatedStorage: aws.Int32(int32(42))}
 	rdsInstance.PendingModifiedValues = &pendingModifications
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, true, metrics.Instances[*rdsInstance.DBInstanceIdentifier].PendingModifiedValues, "Should have allocated storage pending modification")
@@ -351,14 +270,14 @@ func TestPendingModificationDueToInstanceModification(t *testing.T) {
 
 func TestPendingModificationDueToUnappliedParameterGroup(t *testing.T) {
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	rdsInstance.DBParameterGroups = []aws_rds_types.DBParameterGroupStatus{{DBParameterGroupName: aws.String("my_parameter_group"), ParameterApplyStatus: aws.String("pending-reboot")}}
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	require.NoError(t, err, "GetInstancesMetrics must succeed")
 	assert.Equal(t, true, metrics.Instances[*rdsInstance.DBInstanceIdentifier].PendingModifiedValues, "Should have pending modification")
@@ -366,15 +285,15 @@ func TestPendingModificationDueToUnappliedParameterGroup(t *testing.T) {
 
 func TestInstanceAge(t *testing.T) {
 	// Mock RDS instance
-	rdsInstance := newRdsInstance()
+	rdsInstance := mock.NewRdsInstance()
 	creationDate := time.Date(2023, 9, 25, 12, 25, 0, 0, time.UTC) // Date of our first release
 	rdsInstance.InstanceCreateTime = &creationDate
 	mockDescribeDBInstancesOutput := &aws_rds.DescribeDBInstancesOutput{DBInstances: []aws_rds_types.DBInstance{*rdsInstance}}
 
-	mock := mockRDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
+	client := mock.RDSClient{DescribeDBInstancesOutput: mockDescribeDBInstancesOutput}
 	configuration := rds.Configuration{}
-	client := rds.NewFetcher(mock, configuration)
-	metrics, err := client.GetInstancesMetrics()
+	fetcher := rds.NewFetcher(client, configuration)
+	metrics, err := fetcher.GetInstancesMetrics()
 
 	expectedAge := time.Since(*rdsInstance.InstanceCreateTime)
